move commands to thirdparty/commands
TODO remove references to IPFS

cc @mappum @jbenet
