+++ core/commands/commands/commands.go
@@ -5,7 +5,7 @@ import (
 	"io"
 	"sort"
 
-	cmds "github.com/jbenet/go-ipfs/commands"
+	cmds "github.com/jbenet/go-ipfs/thirdparty/commands"
 )
 
 type Command struct {
