+++ core/commands/diag/diag.go
@@ -8,7 +8,7 @@ import (
 	"text/template"
 	"time"
 
-	cmds "github.com/jbenet/go-ipfs/commands"
+	cmds "github.com/jbenet/go-ipfs/thirdparty/commands"
 	ccutil "github.com/jbenet/go-ipfs/core/commands/util"
 	diag "github.com/jbenet/go-ipfs/diagnostics"
 )
