+++ core/commands/name/publish.go
@@ -8,7 +8,7 @@ import (
 
 	b58 "github.com/jbenet/go-ipfs/Godeps/_workspace/src/github.com/jbenet/go-base58"
 
-	cmds "github.com/jbenet/go-ipfs/commands"
+	cmds "github.com/jbenet/go-ipfs/thirdparty/commands"
 	core "github.com/jbenet/go-ipfs/core"
 	nsys "github.com/jbenet/go-ipfs/namesys"
 	crypto "github.com/jbenet/go-ipfs/p2p/crypto"
