+++ core/commands/repo/repo.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"io"
 
-	cmds "github.com/jbenet/go-ipfs/commands"
+	cmds "github.com/jbenet/go-ipfs/thirdparty/commands"
 	corerepo "github.com/jbenet/go-ipfs/core/corerepo"
 	u "github.com/jbenet/go-ipfs/util"
 )
