+++ core/commands/version/version.go
@@ -5,7 +5,7 @@ import (
 	"io"
 	"strings"
 
-	cmds "github.com/jbenet/go-ipfs/commands"
+	cmds "github.com/jbenet/go-ipfs/thirdparty/commands"
 	config "github.com/jbenet/go-ipfs/repo/config"
 )
